// src/types/event.ts
export type TEvent = {
  id: string
  title: string
  description: string
  resourceURI: string
  urls: any
  modified: string
  start: string
  end: string
  thumbnail: {
    path: string
    extension: string
  }
  comics: any
  stories: any
  series: any
  characters: any
  creators: any
  next: any
  previous: any
}